messaging_app/requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
django-filter>=23.2
mysqlclient==2.2.7
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
djangorestframework-simplejwt==5.5.1
python-decouple>=3.8
django-environ==0.9.0
drf-nested-routers==0.93.3
